COCOM coordinates
Out of box, Tableau provides great funcitonality for custom mapping regions as long as your provide your own coordinates. In this case, if you need a shape for a US military command on a map, you don't have an easy way to add that shape in a place that you can easily access and amend as needed. This commit fixes that by providing a data source to blend into a map viz to overlay shapes in a way that makes sense to the viewer.

To use: add as a data source to Tableau and join to a dataset that contains data intended for mapping. Use the provided Lat and Long measures only, not Tableau's auto-generated pills.

"nearest ocean" means your Tableau shape appers over the closest body of water instead of on the actual land mass.